Solved: make Foliage exit when user quits taskbar widget
If the user right-clicks on the taskbar widget and chooses exit,
previously it would exit the taskbar widget but the main application
would keep running.  I finally solved how to make the main application
exit if the widget exits.  It required using a polling approach
instead of a pure event-based wait on PyWebIO pins, but with a 1-sec
polling interval, hopefully the impact on system resources will be
minimal.