Be careful about -p arg
Foliage takes a command line flag -p for port, and it's meant to be
used like -p 8080, but apparently, if you hand it something like
-psn_0_349807 then plac will still return a value (specifically
"sn_0_349807").  That's not supposed to happen because in plac you
tell it what the allowed single-letter and multi-letter options are,
separately.  Anyway, when running on Catalina, in the
PyInstaller-built app, Foliage was getting passed such an argument and
it caused the port-number-checking code to treat it as an error.