Fix handling of missing SRS records
Phuong reported an error involving deletion of an instance, and the
problem turned out to be that there was no corresponding SRS record
for the instance.  Apparently this is not necessarily an error,
depending on the nature of the instance record.  I changed the
handling so that a missing SRS record is not cause to prevent
continuing with the deletion of the instance record in the inventory
system.