fix: _version value needs to be updated
Laurel reported errors started happening. The error was of the form

```
Error: FOLIO server error: ERROR: Cannot update record
6b01869d-b1a5-4157-b27a-c80767edcb7d because it has been changed
(optimistic locking): Stored _version is 3, _version of request is 2
(23F09)
```

My best guess is that some change on the Folio/Ebsco server side
caused this, and now the `_version` value of records needs to be
incremented when making changes to records. So that's what I
implemented here.